Add configs widgets tests.
